Remove outdated DataRepository, and fix location of files for unit tests
