+++ src/main/java/com/dlsc/jfxcentral/data/DataRepository.java
@@ -34,12 +34,8 @@ import com.rometools.rome.feed.synd.SyndEntry;
 import com.rometools.rome.feed.synd.SyndFeed;
 import com.rometools.rome.io.SyndFeedInput;
 import com.rometools.rome.io.XmlReader;
-import javafx.beans.property.ObjectProperty;
-import javafx.beans.property.SimpleObjectProperty;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
 import org.apache.commons.lang3.StringUtils;
 
 import java.io.BufferedReader;
@@ -47,21 +43,15 @@ import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
-import java.net.HttpURLConnection;
-import java.net.MalformedURLException;
-import java.net.ProtocolException;
-import java.net.URL;
-import java.net.URLConnection;
+import java.lang.reflect.Type;
+import java.net.*;
 import java.nio.charset.StandardCharsets;
 import java.text.MessageFormat;
-import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.Collections;
-import java.util.Comparator;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import java.util.logging.Logger;
 import java.util.stream.Collectors;
@@ -69,50 +59,39 @@ import java.util.stream.Collectors;
 public class DataRepository {
 
     private static final Logger LOG = Logger.getLogger(DataRepository.class.getName());
-
     public static File REPO_DIRECTORY = new File(System.getProperty("jfxcentral.repo", new File(System.getProperty("user.home"), ".jfxcentralrepo").getAbsolutePath())).getAbsoluteFile();
-
     private static final DateTimeFormatter DATE_FORMATTER = DateTimeFormatter.ofPattern("yyyy-MM-dd");
-
+    private static boolean testing;
     private static DataRepository instance;
-
     private final Gson gson = Converters.registerLocalDate(new GsonBuilder()).setPrettyPrinting().create();
+    private final List<Library> libraries = new ArrayList<>();
+    private final List<Blog> blogs = new ArrayList<>();
+    private final List<News> news = new ArrayList<>();
+    private final List<Book> books = new ArrayList<>();
+    private final List<LinksOfTheWeek> linksOfTheWeek = new ArrayList<>();
+    private final List<Tip> tips = new ArrayList<>();
+    private final List<Tutorial> tutorials = new ArrayList<>();
+    private final List<Video> videos = new ArrayList<>();
+    private final List<Download> downloads = new ArrayList<>();
+    private final List<RealWorldApp> realWorldApps = new ArrayList<>();
+    private final List<Tool> tools = new ArrayList<>();
+    private final List<Utility> utilities = new ArrayList<>();
+    private final List<Company> companies = new ArrayList<>();
+    private final List<Person> people = new ArrayList<>();
+    private final List<IkonliPack> ikonliPacks = new ArrayList<>();
+    private final List<Member> members = new ArrayList<>();
+    private final List<Documentation> documentation = new ArrayList<>();
+    private final List<LearnJavaFX> learnJavaFX = new ArrayList<>();
+    private final List<LearnMobile> learnMobile = new ArrayList<>();
+    private final List<LearnRaspberryPi> learnRaspberryPi = new ArrayList<>();
+    private String homeText;
+    private String openJFXText;
+    private long cachedPullRequestsTime;
+    private List<PullRequest> cachedPullRequests;
 
-    private final Map<Library, ObjectProperty<LibraryInfo>> libraryInfoMap = new HashMap<>();
-
-    private final Map<News, StringProperty> newsTextMap = new HashMap<>();
-
-    private final Map<Tutorial, StringProperty> tutorialTextMap = new HashMap<>();
-
-    private final Map<Download, StringProperty> downloadTextMap = new HashMap<>();
-
-    private final Map<Book, StringProperty> bookTextMap = new HashMap<>();
-
-    private final Map<Person, StringProperty> personDescriptionMap = new HashMap<>();
-
-    private final Map<Tool, StringProperty> toolDescriptionMap = new HashMap<>();
-
-    private final Map<Utility, StringProperty> utilityDescriptionMap = new HashMap<>();
-
-    private final Map<Tip, StringProperty> tipDescriptionMap = new HashMap<>();
-
-    private final Map<RealWorldApp, StringProperty> realWorldAppDescriptionMap = new HashMap<>();
-
-    private final Map<Company, StringProperty> companyDescriptionMap = new HashMap<>();
-
-    private final Map<Library, StringProperty> libraryReadMeMap = new HashMap<>();
-
-    private final Map<LinksOfTheWeek, StringProperty> linksOfTheWeekReadMeMap = new HashMap<>();
-
-    private final Map<Member, StringProperty> memberDescriptionMap = new HashMap<>();
-    //private final Map<Learn, StringProperty> learnDescriptionMap = new HashMap<>();
-    private final Map<LearnJavaFX,StringProperty> learnJavaFXDescriptionMap = new HashMap<>();
-    private final Map<LearnMobile,StringProperty> learnMobileDescriptionMap = new HashMap<>();
-    private final Map<LearnRaspberryPi,StringProperty> learnRaspberryPiDescriptionMap = new HashMap<>();
-
-    private boolean loaded;
-
-    public static boolean testing = false;
+    private DataRepository() {
+        doLoadData("initial loading of data upon creation of data repository instance");
+    }
 
     public static synchronized DataRepository getInstance() {
         if (instance == null) {
@@ -122,41 +101,23 @@ public class DataRepository {
         return instance;
     }
 
-    public boolean isLoaded() {
-        return loaded;
+    public static void setTesting(boolean testing) {
+        DataRepository.testing = testing;
     }
 
-    private DataRepository() {
+    public static boolean isTesting() {
+        return DataRepository.testing;
     }
 
-    public void loadData() {
+    public void reload() {
         doLoadData("explicit call to refresh method");
     }
 
     public void clearData() {
-        loaded = false;
+        LOG.fine("clearing data");
 
-        setHomeText("");
-        setOpenJFXText("");
-
-        libraryInfoMap.clear();
-        libraryReadMeMap.clear();
-        newsTextMap.clear();
-        personDescriptionMap.clear();
-        companyDescriptionMap.clear();
-        toolDescriptionMap.clear();
-        utilityDescriptionMap.clear();
-        tipDescriptionMap.clear();
-        realWorldAppDescriptionMap.clear();
-        downloadTextMap.clear();
-        bookTextMap.clear();
-        tutorialTextMap.clear();
-        linksOfTheWeekReadMeMap.clear();
-        memberDescriptionMap.clear();
-        learnJavaFXDescriptionMap.clear();
-        learnMobileDescriptionMap.clear();
-        learnRaspberryPiDescriptionMap.clear();
-
+        homeText = "";
+        openJFXText = "";
         getPeople().clear();
         getLibraries().clear();
         getBooks().clear();
@@ -180,162 +141,74 @@ public class DataRepository {
     }
 
     private void doLoadData(String reason) {
+        clearData();
+
         LOG.fine("loading data, reason = " + reason);
 
         try {
-            String homeText = loadString(new File(getRepositoryDirectory(), "intro.md"));
+            // plain texts
+            homeText = loadString(new File(getRepositoryDirectory(), "intro.md"));
+            openJFXText = loadString(new File(getRepositoryDirectory(), "openjfx/intro.md"));
 
-            String openJFXText = loadString(new File(getRepositoryDirectory(), "openjfx/intro.md"));
+            // collections
+            people.addAll(load(getFile("people/people.json"), new TypeToken<List<Person>>() {
+            }.getType()));
+            books.addAll(load(getFile("books/books.json"), new TypeToken<List<Book>>() {
+            }.getType()));
+            videos.addAll(load(getFile("videos/videos.json"), new TypeToken<List<Video>>() {
+            }.getType()));
+            libraries.addAll(load(getFile("libraries/libraries.json"), new TypeToken<List<Library>>() {
+            }.getType()));
+            news.addAll(load(getFile("news/news.json"), new TypeToken<List<News>>() {
+            }.getType()));
+            tutorials.addAll(load(getFile("tutorials/tutorials.json"), new TypeToken<List<Tutorial>>() {
+            }.getType()));
+            blogs.addAll(load(getFile("blogs/blogs.json"), new TypeToken<List<Blog>>() {
+            }.getType()));
+            companies.addAll(load(getFile("companies/companies.json"), new TypeToken<List<Company>>() {
+            }.getType()));
+            tools.addAll(load(getFile("tools/tools.json"), new TypeToken<List<Tool>>() {
+            }.getType()));
+            utilities.addAll(load(getFile("utilities/utilities.json"), new TypeToken<List<Utility>>() {
+            }.getType()));
+            realWorldApps.addAll(load(getFile("realworld/realworld.json"), new TypeToken<List<RealWorldApp>>() {
+            }.getType()));
+            downloads.addAll(load(getFile("downloads/downloads.json"), new TypeToken<List<Download>>() {
+            }.getType()));
+            tips.addAll(load(getFile("tips/tips.json"), new TypeToken<List<Tip>>() {
+            }.getType()));
+            linksOfTheWeek.addAll(load(getFile("links/links.json"), new TypeToken<List<LinksOfTheWeek>>() {
+            }.getType()));
+            ikonliPacks.addAll(load(getFile("ikonlipacks/ikonlipacks.json"), new TypeToken<List<IkonliPack>>() {
+            }.getType()));
+            members.addAll(load(getFile("members/members.json"), new TypeToken<List<Member>>() {
+            }.getType()));
+            documentation.addAll(load(getFile("documentation/documentation.json"), new TypeToken<List<Documentation>>() {
+            }.getType()));
+            learnJavaFX.addAll(load(getFile("learn/javafx/learn.json"), new TypeToken<List<LearnJavaFX>>() {
+            }.getType()));
+            learnMobile.addAll(load(getFile("learn/mobile/learn.json"), new TypeToken<List<LearnMobile>>() {
+            }.getType()));
+            learnRaspberryPi.addAll(load(getFile("learn/raspberrypi/learn.json"), new TypeToken<List<LearnRaspberryPi>>() {
+            }.getType()));
 
-            List<Person> people = loadModel("people/people.json", new TypeToken<>() {
-            });
-            List<Book> books = loadModel("books/books.json", new TypeToken<>() {
-            });
-            List<Video> videos = loadModel("videos/videos.json", new TypeToken<>() {
-            });
-            List<Library> libraries = loadModel("libraries/libraries.json", new TypeToken<>() {
-            });
-            List<News> news = loadModel("news/news.json", new TypeToken<>() {
-            });
-            List<Blog> blogs = loadModel("blogs/blogs.json", new TypeToken<>() {
-            });
-            List<Company> companies = loadModel("companies/companies.json", new TypeToken<>() {
-            });
-            List<Tool> tools = loadModel("tools/tools.json", new TypeToken<>() {
-            });
-            List<Utility> utilities = loadModel("utilities/utilities.json", new TypeToken<>() {
-            });
-            List<RealWorldApp> realWorldApps = loadModel("realworld/realworld.json", new TypeToken<>() {
-            });
-            List<Download> downloads = loadModel("downloads/downloads.json", new TypeToken<>() {
-            });
-            List<Tutorial> tutorials = loadModel("tutorials/tutorials.json", new TypeToken<>() {
-            });
-            List<Tip> tips = loadModel("tips/tips.json", new TypeToken<>() {
-            });
-            List<LinksOfTheWeek> links = loadModel("links/links.json", new TypeToken<>() {
-            });
-            List<IkonliPack> ikonliPacks = loadModel("ikonlipacks/ikonlipacks.json", new TypeToken<>() {
-            });
-            List<Member> members = loadModel("members/members.json", new TypeToken<>() {
-            });
-            List<Documentation> documentation = loadModel("documentation/documentation.json", new TypeToken<>() {
-            });
-            List<LearnJavaFX> learnJavaFX = loadModel("learn/javafx/learn.json", new TypeToken<>() {
-            });
-            List<LearnMobile> learnMobile = loadModel("learn/mobile/learn.json", new TypeToken<>() {
-            });
-            List<LearnRaspberryPi> learnRaspberryPi = loadModel("learn/raspberrypi/learn.json", new TypeToken<>() {
-            });
-
-            setData(homeText, openJFXText, people, books, videos, libraries, news, blogs, companies, tools, utilities, realWorldApps, downloads, tutorials, tips, links, ikonliPacks, members, documentation, learnJavaFX, learnMobile, learnRaspberryPi);
-
-            LOG.fine("data loading finished");
         } catch (Exception e) {
-            LOG.severe("data loading failed: " + e.getMessage());
-        } finally {
-            loaded = true;
+            LOG.severe("error loading data: " + e.getMessage());
         }
+
+        LOG.fine("data loading finished");
     }
 
-    private <T> List<T> loadModel(String path, TypeToken<List<T>> typeToken) {
-        File file = new File(getRepositoryDirectory(), path);
-        try (FileReader fileReader = new FileReader(file, StandardCharsets.UTF_8)) {
-            return gson.fromJson(fileReader, typeToken.getType());
+    private <T> Collection<T> load(File file, Type type) {
+        try (FileReader fr = new FileReader(file, StandardCharsets.UTF_8)) {
+            return gson.fromJson(fr, type);
         } catch (IOException e) {
-            LOG.severe("failed to load model from file " + path + ": " + e.getMessage());
-            return Collections.emptyList();
+            throw new RuntimeException(e);
         }
     }
 
-    private void setData(String homeText, String openJFXText, List<Person> people, List<Book> books, List<Video> videos, List<Library> libraries,
-                         List<News> news, List<Blog> blogs, List<Company> companies, List<Tool> tools, List<Utility> utilities, List<RealWorldApp> realWorldApps, List<Download> downloads,
-                         List<Tutorial> tutorials, List<Tip> tips, List<LinksOfTheWeek> links, List<IkonliPack> ikonliPacks, List<Member> members, List<Documentation> documentation, List<LearnJavaFX> learnJavaFX, List<LearnMobile> learnMobile, List<LearnRaspberryPi> learnRaspberryPi) {
-        clearData();
-
-        setOpenJFXText(openJFXText);
-        setHomeText(homeText);
-
-        getPeople().setAll(people);
-        getBooks().setAll(books);
-        getVideos().setAll(videos);
-        getLibraries().setAll(libraries);
-        getNews().setAll(news);
-        getBlogs().setAll(blogs);
-        getCompanies().setAll(companies);
-        getTools().setAll(tools);
-        getUtilities().setAll(utilities);
-        getRealWorldApps().setAll(realWorldApps);
-        getDownloads().setAll(downloads);
-        getTutorials().setAll(tutorials);
-        getTips().setAll(tips);
-        getLinksOfTheWeek().setAll(links);
-        getIkonliPacks().setAll(ikonliPacks);
-        getMembers().setAll(members);
-        getDocumentation().setAll(documentation);
-        getLearnJavaFX().setAll(learnJavaFX);
-        getLearnMobile().setAll(learnMobile);
-        getLearnRaspberryPi().setAll(learnRaspberryPi);
-
-        List<ModelObject> recentItems = findRecentItems();
-        getRecentItems().setAll(recentItems);
-    }
-
-    private List<ModelObject> findRecentItems() {
-        List<ModelObject> result = new ArrayList<>();
-        // News are not reachable through links!
-        //  result.addAll(findRecentItems(getNews()));
-        result.addAll(findRecentItems(getPeople()));
-        result.addAll(findRecentItems(getBooks()));
-        result.addAll(findRecentItems(getLibraries()));
-        result.addAll(findRecentItems(getVideos()));
-        result.addAll(findRecentItems(getBlogs()));
-        result.addAll(findRecentItems(getCompanies()));
-        result.addAll(findRecentItems(getTools()));
-        result.addAll(findRecentItems(getUtilities()));
-        result.addAll(findRecentItems(getTutorials()));
-        result.addAll(findRecentItems(getRealWorldApps()));
-        result.addAll(findRecentItems(getDownloads()));
-        result.addAll(findRecentItems(getTips()));
-        result.addAll(findRecentItems(getIkonliPacks()));
-        result.addAll(findRecentItems(getMembers()));
-        result.addAll(findRecentItems(getDocumentation()));
-        result.addAll(findRecentItems(getLearnJavaFX()));
-        result.addAll(findRecentItems(getLearnMobile()));
-        result.addAll(findRecentItems(getLearnRaspberryPi()));
-        // LinksOfTheWeek are not reachable through links!
-        //  result.addAll(findRecentItems(getLinksOfTheWeek()));
-
-        // newest ones on top
-        Collections.sort(result, Comparator.comparing(ModelObject::getCreationOrUpdateDate).reversed());
-
-        return result;
-    }
-
-    private List<ModelObject> findRecentItems(List<? extends ModelObject> items) {
-        List<ModelObject> result = new ArrayList<>();
-
-        final LocalDate today = LocalDate.now();
-
-        items.forEach(item -> {
-            LocalDate date = item.getModifiedOn();
-            if (date == null) {
-                date = item.getCreatedOn();
-            }
-            if (date != null) {
-                if (date.isAfter(today.minusWeeks(8))) {
-                    result.add(item);
-                }
-            }
-        });
-
-        return result;
-    }
-
-    private final ObservableList<ModelObject> recentItems = FXCollections.observableArrayList();
-
-    public ObservableList<ModelObject> getRecentItems() {
-        return recentItems;
+    private File getFile(String path) {
+        return new File(getRepositoryDirectory(), path);
     }
 
     public Optional<Person> getPersonById(String id) {
@@ -352,6 +225,18 @@ public class DataRepository {
 
     public Optional<Book> getBookById(String id) {
         return books.stream().filter(item -> item.getId().equals(id)).findFirst();
+    }
+
+    public Optional<LearnJavaFX> getLearnJavaFXById(String id) {
+        return learnJavaFX.stream().filter(item -> item.getId().equals(id)).findFirst();
+    }
+
+    public Optional<LearnMobile> getLearnMobileById(String id) {
+        return learnMobile.stream().filter(item -> item.getId().equals(id)).findFirst();
+    }
+
+    public Optional<LearnRaspberryPi> getLearnRaspberryPiById(String id) {
+        return learnRaspberryPi.stream().filter(item -> item.getId().equals(id)).findFirst();
     }
 
     public Optional<Blog> getBlogById(String id) {
@@ -406,24 +291,10 @@ public class DataRepository {
         return documentation.stream().filter(item -> item.getId().equals(id)).findFirst();
     }
 
-    public Optional<LearnJavaFX> getLearnJavaFXById(String id) {
-        return learnJavaFX.stream().filter(item -> item.getId().equals(id)).findFirst();
-    }
-
-    public Optional<LearnMobile> getLearnMobileById(String id) {
-        return learnMobile.stream().filter(item -> item.getId().equals(id)).findFirst();
-    }
-
-    public Optional<LearnRaspberryPi> getLearnRaspberryPiById(String id) {
-        return learnRaspberryPi.stream().filter(item -> item.getId().equals(id)).findFirst();
-    }
-
-    public <T extends ModelObject> ObservableList<T> getLinkedObjects(ModelObject modelObject, Class<T> clazz) {
+    public <T extends ModelObject> List<T> getLinkedObjects(ModelObject modelObject, Class<T> clazz) {
         List<T> itemList = getList(clazz);
         List<String> idsList = getIdList(modelObject, clazz);
-        ObservableList<T> list = FXCollections.observableArrayList();
-        list.setAll(itemList.stream().filter(item -> idsList.contains(item.getId()) || getIdList(item, modelObject.getClass()).contains(modelObject.getId())).collect(Collectors.toList()));
-        return list;
+        return itemList.stream().filter(item -> idsList.contains(item.getId()) || getIdList(item, modelObject.getClass()).contains(modelObject.getId())).collect(Collectors.toList());
     }
 
     private <T extends ModelObject> List<String> getIdList(ModelObject modelObject, Class<T> clazz) {
@@ -518,468 +389,257 @@ public class DataRepository {
         throw new IllegalArgumentException("unsupported class type: " + clazz.getSimpleName());
     }
 
-    public ModelObject getByID(Class<? extends ModelObject> clz, String id) {
-        return getList(clz).stream().filter(item -> item.getId().equals(id)).findFirst().get();
+    public <T extends ModelObject> T getByID(Class<T> clz, String id) {
+        return getList(clz).stream().filter(item -> item.getId().equals(id)).findFirst().orElse(null);
     }
 
-    public ObservableList<Video> getVideosByModelObject(ModelObject modelObject) {
+    public boolean isValidId(Class<? extends ModelObject> clz, String id) {
+        return getList(clz).stream().anyMatch(item -> item.getId().equals(id));
+    }
+
+    public List<Video> getVideosByModelObject(ModelObject modelObject) {
         return getLinkedObjects(modelObject, Video.class);
     }
 
-    public ObservableList<Download> getDownloadsByModelObject(ModelObject modelObject) {
+    public List<Download> getDownloadsByModelObject(ModelObject modelObject) {
         return getLinkedObjects(modelObject, Download.class);
     }
 
-    public ObservableList<LearnJavaFX> getLearnJavaFXByModelObject(ModelObject modelObject) {
-        return getLinkedObjects(modelObject, LearnJavaFX.class);
-    }
-
-    public ObservableList<LearnMobile> getLearnMobileByModelObject(ModelObject modelObject) {
-        return getLinkedObjects(modelObject, LearnMobile.class);
-    }
-
-    public ObservableList<LearnRaspberryPi> getLearnRaspberryPiByModelObject(ModelObject modelObject) {
-        return getLinkedObjects(modelObject, LearnRaspberryPi.class);
-    }
-
-    public ObservableList<Book> getBooksByModelObject(ModelObject modelObject) {
+    public List<Book> getBooksByModelObject(ModelObject modelObject) {
         return getLinkedObjects(modelObject, Book.class);
     }
 
-    public ObservableList<Tutorial> getTutorialsByModelObject(ModelObject modelObject) {
+    public List<Tutorial> getTutorialsByModelObject(ModelObject modelObject) {
         return getLinkedObjects(modelObject, Tutorial.class);
     }
 
-    public ObservableList<Blog> getBlogsByModelObject(ModelObject modelObject) {
+    public List<Blog> getBlogsByModelObject(ModelObject modelObject) {
         return getLinkedObjects(modelObject, Blog.class);
     }
 
-    public ObservableList<Library> getLibrariesByModelObject(ModelObject modelObject) {
+    public List<Library> getLibrariesByModelObject(ModelObject modelObject) {
         return getLinkedObjects(modelObject, Library.class);
     }
 
-    public ObservableList<Tool> getToolsByModelObject(ModelObject modelObject) {
+    public List<Tool> getToolsByModelObject(ModelObject modelObject) {
         return getLinkedObjects(modelObject, Tool.class);
     }
 
-    public ObservableList<News> getNewsByModelObject(ModelObject modelObject) {
+    public List<Utility> getUtilitiesByModelObject(ModelObject modelObject) {
+        return getLinkedObjects(modelObject, Utility.class);
+    }
+
+    public List<News> getNewsByModelObject(ModelObject modelObject) {
         return getLinkedObjects(modelObject, News.class);
     }
 
-    public ObservableList<Company> getCompaniesByModelObject(ModelObject modelObject) {
+    public List<Company> getCompaniesByModelObject(ModelObject modelObject) {
         return getLinkedObjects(modelObject, Company.class);
     }
 
-    public ObservableList<RealWorldApp> getRealWorldAppsByModelObject(ModelObject modelObject) {
+    public List<RealWorldApp> getRealWorldAppsByModelObject(ModelObject modelObject) {
         return getLinkedObjects(modelObject, RealWorldApp.class);
     }
 
-    public ObservableList<Person> getPeopleByModelObject(ModelObject modelObject) {
+    public List<Person> getPeopleByModelObject(ModelObject modelObject) {
         return getLinkedObjects(modelObject, Person.class);
     }
 
-    public ObservableList<Tip> getTipsByModelObject(ModelObject modelObject) {
+    public List<Tip> getTipsByModelObject(ModelObject modelObject) {
         return getLinkedObjects(modelObject, Tip.class);
     }
 
-    public ObservableList<LinksOfTheWeek> getLinksOfTheWeekByModelObject(ModelObject modelObject) {
+    public List<LearnJavaFX> getLearnJavaFXByModelObject(ModelObject modelObject) {
+        return getLinkedObjects(modelObject, LearnJavaFX.class);
+    }
+
+    public List<LearnMobile> getLearnMobileByModelObject(ModelObject modelObject) {
+        return getLinkedObjects(modelObject, LearnMobile.class);
+    }
+
+    public List<LearnRaspberryPi> getLearnRaspberryPiByModelObject(ModelObject modelObject) {
+        return getLinkedObjects(modelObject, LearnRaspberryPi.class);
+    }
+
+    public List<LinksOfTheWeek> getLinksOfTheWeekByModelObject(ModelObject modelObject) {
         return getLinkedObjects(modelObject, LinksOfTheWeek.class);
     }
 
-    public ObjectProperty<LibraryInfo> libraryInfoProperty(Library library) {
-        return libraryInfoMap.computeIfAbsent(library, key -> {
-            ObjectProperty<LibraryInfo> infoProperty = new SimpleObjectProperty<>();
-            loadLibraryInfo(library, infoProperty);
-            return infoProperty;
-        });
-    }
-
-    private void loadLibraryInfo(Library library, ObjectProperty<LibraryInfo> infoProperty) {
+    public LibraryInfo getLibraryInfo(Library library) {
         try {
             String libraryId = library.getId();
             File file = new File(getRepositoryDirectory(), "libraries/" + libraryId + "/info.json");
             try (FileReader reader = new FileReader(file, StandardCharsets.UTF_8)) {
-                LibraryInfo result = gson.fromJson(reader, LibraryInfo.class);
-                infoProperty.set(result);
+                return gson.fromJson(reader, LibraryInfo.class);
             }
         } catch (IOException ex) {
             ex.printStackTrace();
         }
-    }
-
-    public StringProperty newsTextProperty(News news) {
-        return newsTextMap.computeIfAbsent(news, key -> {
-            StringProperty textProperty = new SimpleStringProperty();
-            loadNewsText(news, textProperty);
-            return textProperty;
-        });
-    }
-
-    private void loadNewsText(News news, StringProperty textProperty) {
-        String text = loadString(new File(getNewsDirectory(news), "/text.md"));
-        textProperty.set(text);
-    }
-
-    public StringProperty linksOfTheWeekTextProperty(LinksOfTheWeek links) {
-        return linksOfTheWeekReadMeMap.computeIfAbsent(links, key -> {
-            StringProperty textProperty = new SimpleStringProperty();
-            loadLinksOfTheWeekText(links, textProperty);
-            return textProperty;
-        });
-    }
-
-    public void loadLinksOfTheWeekText(LinksOfTheWeek links, StringProperty textProperty) {
-        String text = loadString(new File(getRepositoryDirectory(), "links/" + links.getId() + "/readme.md"));
-        textProperty.set(text);
-    }
-
-    public StringProperty tutorialTextProperty(Tutorial tutorial) {
-        return tutorialTextMap.computeIfAbsent(tutorial, key -> {
-            StringProperty textProperty = new SimpleStringProperty();
-            loadTutorialText(tutorial, textProperty);
-            return textProperty;
-        });
-    }
-
-    private void loadTutorialText(Tutorial tutorial, StringProperty textProperty) {
-        String text = loadString(new File(getRepositoryDirectory(), "tutorials/" + tutorial.getId() + "/readme.md"));
-        textProperty.set(text);
-    }
-
-    public StringProperty downloadTextProperty(Download download) {
-        return downloadTextMap.computeIfAbsent(download, key -> {
-            StringProperty textProperty = new SimpleStringProperty();
-            loadDownloadText(download, textProperty);
-            return textProperty;
-        });
-    }
-
-    private void loadDownloadText(Download download, StringProperty textProperty) {
-        String text = loadString(new File(getRepositoryDirectory(), "downloads/" + download.getId() + "/readme.md"));
-        textProperty.set(text);
-    }
-
-    public StringProperty bookTextProperty(Book book) {
-        return bookTextMap.computeIfAbsent(book, key -> {
-            StringProperty textProperty = new SimpleStringProperty();
-            loadBookText(book, textProperty);
-            return textProperty;
-        });
-    }
-
-    private void loadBookText(Book book, StringProperty textProperty) {
-        String text = loadString(new File(getRepositoryDirectory(), "books/" + book.getId() + "/readme.md"));
-        textProperty.set(text);
-    }
-
-    public StringProperty personDescriptionProperty(Person person) {
-        return personDescriptionMap.computeIfAbsent(person, key -> {
-            StringProperty readmeProperty = new SimpleStringProperty();
-            loadPersonDescription(person, readmeProperty);
-            return readmeProperty;
-        });
-    }
-
-    private void loadPersonDescription(Person person, StringProperty readmeProperty) {
-        String readmeText = loadString(new File(getRepositoryDirectory(), "people/" + person.getId() + "/readme.md"));
-        readmeProperty.set(readmeText);
-    }
-
-    public StringProperty memberDescriptionProperty(Member member) {
-        return memberDescriptionMap.computeIfAbsent(member, key -> {
-            StringProperty readmeProperty = new SimpleStringProperty();
-            loadMemberDescription(member, readmeProperty);
-            return readmeProperty;
-        });
-    }
-
-    private void loadMemberDescription(Member member, StringProperty readmeProperty) {
-        String readmeText = loadString(new File(getRepositoryDirectory(), "members/" + member.getId() + "/readme.md"));
-        readmeProperty.set(readmeText);
-    }
-
-    public StringProperty toolDescriptionProperty(Tool tool) {
-        return toolDescriptionMap.computeIfAbsent(tool, key -> {
-            StringProperty readmeProperty = new SimpleStringProperty();
-            loadToolDescription(tool, readmeProperty);
-            return readmeProperty;
-        });
-    }
-
-    private void loadToolDescription(Tool tool, StringProperty readmeProperty) {
-        String readmeText = loadString(new File(getRepositoryDirectory(), "tools/" + tool.getId() + "/readme.md"));
-        readmeProperty.set(readmeText);
-    }
-
-    public StringProperty utilityDescriptionProperty(Utility utility) {
-        return utilityDescriptionMap.computeIfAbsent(utility, key -> {
-            StringProperty readmeProperty = new SimpleStringProperty();
-            loadUtilityDescription(utility, readmeProperty);
-            return readmeProperty;
-        });
-    }
-
-    private void loadUtilityDescription(Utility utility, StringProperty readmeProperty) {
-        String readmeText = loadString(new File(getRepositoryDirectory(), "utilities/" + utility.getId() + "/readme.md"));
-        readmeProperty.set(readmeText);
-    }
-
-    public StringProperty tipDescriptionProperty(Tip tip) {
-        return tipDescriptionMap.computeIfAbsent(tip, key -> {
-            StringProperty readmeProperty = new SimpleStringProperty();
-            loadTipDescription(tip, readmeProperty);
-            return readmeProperty;
-        });
-    }
-
-    private void loadTipDescription(Tip tip, StringProperty readmeProperty) {
-        String readmeText = loadString(new File(getRepositoryDirectory(), "tips/" + tip.getId() + "/readme.md"));
-        readmeProperty.set(readmeText);
-    }
-
-    public StringProperty realWorldAppDescriptionProperty(RealWorldApp app) {
-        return realWorldAppDescriptionMap.computeIfAbsent(app, key -> {
-            StringProperty readmeProperty = new SimpleStringProperty();
-            loadRealWorldDescription(app, readmeProperty);
-            return readmeProperty;
-        });
-    }
-
-    private void loadRealWorldDescription(RealWorldApp app, StringProperty readmeProperty) {
-        String readmeText = loadString(new File(getRepositoryDirectory(), "realworld/" + app.getId() + "/readme.md"));
-        readmeProperty.set(readmeText);
-    }
-
-    public StringProperty companyDescriptionProperty(Company company) {
-        return companyDescriptionMap.computeIfAbsent(company, key -> {
-            StringProperty readmeProperty = new SimpleStringProperty();
-            loadCompanyDescription(company, readmeProperty);
-            return readmeProperty;
-        });
-    }
-
-    private void loadCompanyDescription(Company company, StringProperty readmeProperty) {
-        String readmeText = loadString(new File(getRepositoryDirectory(), "companies/" + company.getId() + "/readme.md"));
-        readmeProperty.set(readmeText);
-    }
-
-    private void loadLearnJavaFXReadMe(LearnJavaFX learnJavaFX, StringProperty readmeProperty) {
-        String readmeText = loadString(new File(getRepositoryDirectory(), "learn/javafx/" + learnJavaFX.getId() + "/readme.md"));
-        readmeProperty.set(readmeText);
-    }
-
-    public StringProperty learnJavaFXReadMeProperty(LearnJavaFX learnJavaFX) {
-        return learnJavaFXDescriptionMap.computeIfAbsent(learnJavaFX, key -> {
-            StringProperty readmeProperty = new SimpleStringProperty();
-            loadLearnJavaFXReadMe(learnJavaFX, readmeProperty);
-            return readmeProperty;
-        });
-    }
-
-    private void loadLearnMobileReadMe(LearnMobile learnMobile, StringProperty readmeProperty) {
-        String readmeText = loadString(new File(getRepositoryDirectory(), "learn/mobile/" + learnMobile.getId() + "/readme.md"));
-        readmeProperty.set(readmeText);
-    }
-
-    public StringProperty learnMobileReadMeProperty(LearnMobile learnMobile) {
-        return learnMobileDescriptionMap.computeIfAbsent(learnMobile, key -> {
-            StringProperty readmeProperty = new SimpleStringProperty();
-            loadLearnMobileReadMe(learnMobile, readmeProperty);
-            return readmeProperty;
-        });
-    }
-
-    private void loadLearnRaspberryPiReadMe(LearnRaspberryPi learnRaspberryPi, StringProperty readmeProperty) {
-        String readmeText = loadString(new File(getRepositoryDirectory(), "learn/raspberrypi/" + learnRaspberryPi.getId() + "/readme.md"));
-        readmeProperty.set(readmeText);
-    }
-
-    public StringProperty learnRaspberryPiReadMeProperty(LearnRaspberryPi learnRaspberryPi) {
-        return learnRaspberryPiDescriptionMap.computeIfAbsent(learnRaspberryPi, key -> {
-            StringProperty readmeProperty = new SimpleStringProperty();
-            loadLearnRaspberryPiReadMe(learnRaspberryPi, readmeProperty);
-            return readmeProperty;
-        });
-    }
-
-    public static void setTesting(boolean testing) {
-        DataRepository.testing = testing;
-    }
-
-    public File getRepositoryDirectory() {
-        if (testing) {
-            return new File(System.getProperty("user.dir"));
-        }
-        return REPO_DIRECTORY;
-    }
-
-    public String getRepositoryDirectoryURL() {
-        return getRepositoryDirectory().toURI().toString();
+        return null;
     }
 
     public File getNewsDirectory(News news) {
         return new File(getRepositoryDirectory(), "news/" + DATE_FORMATTER.format(news.getCreatedOn()) + "-" + news.getId());
     }
 
-    public StringProperty libraryReadMeProperty(Library library) {
-        return libraryReadMeMap.computeIfAbsent(library, key -> {
-            StringProperty readmeProperty = new SimpleStringProperty();
-            loadLibraryDescription(library, readmeProperty);
-            return readmeProperty;
-        });
+    public String getNewsText(News news) {
+        return loadString(new File(getNewsDirectory(news), "/text.md"));
     }
 
-    private void loadLibraryDescription(Library library, StringProperty readmeProperty) {
-        String readmeText = loadString(new File(getRepositoryDirectory(), "libraries/" + library.getId() + "/readme.md"));
-        readmeProperty.set(readmeText);
+    public String getLinksOfTheWeekReadMe(LinksOfTheWeek links) {
+        return loadString(new File(getRepositoryDirectory(), "links/" + links.getId() + "/readme.md"));
     }
 
-    private final StringProperty homeText = new SimpleStringProperty(this, "homeText");
-
-    public String getHomeText() {
-        return homeText.get();
+    public void loadLinksOfTheWeekText(LinksOfTheWeek links, StringProperty textProperty) {
+        textProperty.set(getLinksOfTheWeekReadMe(links));
     }
 
-    public StringProperty homeTextProperty() {
-        return homeText;
+    public String getTutorialReadMe(Tutorial tutorial) {
+        return loadString(new File(getRepositoryDirectory(), "tutorials/" + tutorial.getId() + "/readme.md"));
     }
 
-    public void setHomeText(String homeText) {
-        this.homeText.set(homeText);
+    public String getDownloadReadMe(Download download) {
+        return loadString(new File(getRepositoryDirectory(), "downloads/" + download.getId() + "/readme.md"));
     }
 
-    private final StringProperty openJFXText = new SimpleStringProperty(this, "openJFXText");
-
-    public String getOpenJFXText() {
-        return openJFXText.get();
+    public String getBookReadMe(Book book) {
+        return loadString(new File(getRepositoryDirectory(), "books/" + book.getId() + "/readme.md"));
     }
 
-    public StringProperty openJFXTextProperty() {
-        return openJFXText;
+    public String getPersonReadMe(Person person) {
+        return loadString(new File(getRepositoryDirectory(), "people/" + person.getId() + "/readme.md"));
     }
 
-    public void setOpenJFXText(String openJFXText) {
-        this.openJFXText.set(openJFXText);
+    public String getMemberReadMe(Member member) {
+        return loadString(new File(getRepositoryDirectory(), "members/" + member.getId() + "/readme.md"));
     }
 
-    private final ObservableList<Library> libraries = FXCollections.observableArrayList();
+    public String getToolReadMe(Tool tool) {
+        return loadString(new File(getRepositoryDirectory(), "tools/" + tool.getId() + "/readme.md"));
+    }
 
-    public ObservableList<Library> getLibraries() {
+    public String getUtilityReadMe(Utility utility) {
+        return loadString(new File(getRepositoryDirectory(), "utilities/" + utility.getId() + "/readme.md"));
+    }
+
+    public String getTipReadMe(Tip tip) {
+        return loadString(new File(getRepositoryDirectory(), "tips/" + tip.getId() + "/readme.md"));
+    }
+
+    public String getRealWorldReadMe(RealWorldApp app) {
+        return loadString(new File(getRepositoryDirectory(), "realworld/" + app.getId() + "/readme.md"));
+    }
+
+    public String getCompanyReadMe(Company company) {
+        return loadString(new File(getRepositoryDirectory(), "companies/" + company.getId() + "/readme.md"));
+    }
+
+    public String getLibraryReadMe(Library library) {
+        return loadString(new File(getRepositoryDirectory(), "libraries/" + library.getId() + "/readme.md"));
+    }
+
+    public String getLearnJavaFXReadMe(LearnJavaFX learnJavaFX) {
+        return loadString(new File(getRepositoryDirectory(), "learn/javafx/" + learnJavaFX.getId() + "/readme.md"));
+    }
+
+    public String getLearnMobileReadMe(LearnMobile learnMobile) {
+        return loadString(new File(getRepositoryDirectory(), "learn/mobile/" + learnMobile.getId() + "/readme.md"));
+    }
+
+    public String getLearnRaspberryPiReadMe(LearnRaspberryPi learnRaspberryPi) {
+        return loadString(new File(getRepositoryDirectory(), "learn/raspberrypi/" + learnRaspberryPi.getId() + "/readme.md"));
+    }
+
+    public String getRepositoryDirectoryURL() {
+        return getRepositoryDirectory().toURI().toString();
+    }
+
+    public List<Library> getLibraries() {
         return libraries;
     }
 
-    private final ObservableList<Blog> blogs = FXCollections.observableArrayList();
-
-    public ObservableList<Blog> getBlogs() {
+    public List<Blog> getBlogs() {
         return blogs;
     }
 
-    private final ObservableList<News> news = FXCollections.observableArrayList();
-
-    public ObservableList<News> getNews() {
+    public List<News> getNews() {
         return news;
     }
 
-    private final ObservableList<Book> books = FXCollections.observableArrayList();
-
-    public ObservableList<Book> getBooks() {
+    public List<Book> getBooks() {
         return books;
     }
 
-    private final ObservableList<LinksOfTheWeek> linksOfTheWeek = FXCollections.observableArrayList();
-
-    public ObservableList<LinksOfTheWeek> getLinksOfTheWeek() {
+    public List<LinksOfTheWeek> getLinksOfTheWeek() {
         return linksOfTheWeek;
     }
 
-    private final ObservableList<Tip> tips = FXCollections.observableArrayList();
-
-    public ObservableList<Tip> getTips() {
+    public List<Tip> getTips() {
         return tips;
     }
 
-    private final ObservableList<Tutorial> tutorials = FXCollections.observableArrayList();
-
-    public ObservableList<Tutorial> getTutorials() {
+    public List<Tutorial> getTutorials() {
         return tutorials;
     }
 
-    private final ObservableList<Video> videos = FXCollections.observableArrayList();
-
-    public ObservableList<Video> getVideos() {
+    public List<Video> getVideos() {
         return videos;
     }
 
-    private final ObservableList<Download> downloads = FXCollections.observableArrayList();
-
-    public ObservableList<Download> getDownloads() {
+    public List<Download> getDownloads() {
         return downloads;
     }
 
-    private final ObservableList<RealWorldApp> realWorldApps = FXCollections.observableArrayList();
-
-    public ObservableList<RealWorldApp> getRealWorldApps() {
+    public List<RealWorldApp> getRealWorldApps() {
         return realWorldApps;
     }
 
-    private final ObservableList<Tool> tools = FXCollections.observableArrayList();
-
-    public ObservableList<Tool> getTools() {
+    public List<Tool> getTools() {
         return tools;
     }
 
-    private final ObservableList<Utility> utilities = FXCollections.observableArrayList();
-
-    public ObservableList<Utility> getUtilities() {
+    public List<Utility> getUtilities() {
         return utilities;
     }
 
-    private final ObservableList<Company> companies = FXCollections.observableArrayList();
-
-    public ObservableList<Company> getCompanies() {
+    public List<Company> getCompanies() {
         return companies;
     }
 
-    private final ObservableList<Person> people = FXCollections.observableArrayList();
-
-    public ObservableList<Person> getPeople() {
+    public List<Person> getPeople() {
         return people;
     }
 
-    private final ObservableList<IkonliPack> ikonliPacks = FXCollections.observableArrayList();
-
-    public ObservableList<IkonliPack> getIkonliPacks() {
+    public List<IkonliPack> getIkonliPacks() {
         return ikonliPacks;
     }
 
-    private final ObservableList<Member> members = FXCollections.observableArrayList();
-
-    public ObservableList<Member> getMembers() {
+    public List<Member> getMembers() {
         return members;
     }
 
-    private final ObservableList<Documentation> documentation = FXCollections.observableArrayList();
-
-    public ObservableList<Documentation> getDocumentation() {
+    public List<Documentation> getDocumentation() {
         return documentation;
     }
 
-    private final ObservableList<LearnJavaFX> learnJavaFX = FXCollections.observableArrayList();
-
-    public ObservableList<LearnJavaFX> getLearnJavaFX() {
+    public List<LearnJavaFX> getLearnJavaFX() {
         return learnJavaFX;
     }
 
-    private final ObservableList<LearnMobile> learnMobile = FXCollections.observableArrayList();
-
-    public ObservableList<LearnMobile> getLearnMobile() {
+    public List<LearnMobile> getLearnMobile() {
         return learnMobile;
     }
 
-    private final ObservableList<LearnRaspberryPi> learnRaspberryPi = FXCollections.observableArrayList();
-
-    public ObservableList<LearnRaspberryPi> getLearnRaspberryPi() {
+    public List<LearnRaspberryPi> getLearnRaspberryPi() {
         return learnRaspberryPi;
+    }
+
+    public String getHomeText() {
+        return homeText;
+    }
+
+    public String getOpenJFXText() {
+        return openJFXText;
     }
 
     private String loadString(File file) {
@@ -1030,7 +690,7 @@ public class DataRepository {
             if (status == 200) {
                 try (BufferedReader in = new BufferedReader(new InputStreamReader(con.getInputStream()))) {
                     String inputLine;
-                    StringBuffer content = new StringBuffer();
+                    StringBuilder content = new StringBuilder();
                     while ((inputLine = in.readLine()) != null) {
                         content.append(inputLine);
                     }
@@ -1044,10 +704,6 @@ public class DataRepository {
             } else {
                 result.set("unknown");
             }
-        } catch (ProtocolException e) {
-            e.printStackTrace();
-        } catch (MalformedURLException e) {
-            e.printStackTrace();
         } catch (IOException e) {
             e.printStackTrace();
         } finally {
@@ -1081,18 +737,13 @@ public class DataRepository {
         return Collections.emptyList();
     }
 
-    private long cachedPullrequestsTime;
-
-    private final long timeToReloadSeconds = 600;
-
-    private List<PullRequest> cachedPullRequests;
-
     public List<PullRequest> loadPullRequests() {
         long time = System.currentTimeMillis() / 1000;
-        if (cachedPullrequestsTime + timeToReloadSeconds > time) {
+        long timeToReloadSeconds = 600;
+        if (cachedPullRequestsTime + timeToReloadSeconds > time) {
             return cachedPullRequests;
         }
-        cachedPullrequestsTime = time;
+        cachedPullRequestsTime = time;
         cachedPullRequests = loadPullRequestsImpl();
         return cachedPullRequests;
 
@@ -1115,7 +766,7 @@ public class DataRepository {
                 if (status == 200) {
                     try (BufferedReader in = new BufferedReader(new InputStreamReader(con.getInputStream()))) {
                         String inputLine;
-                        StringBuffer content = new StringBuffer();
+                        StringBuilder content = new StringBuilder();
                         while ((inputLine = in.readLine()) != null) {
                             content.append(inputLine);
                         }
@@ -1123,10 +774,6 @@ public class DataRepository {
                         }.getType());
                     }
                 }
-            } catch (ProtocolException e) {
-                e.printStackTrace();
-            } catch (MalformedURLException e) {
-                e.printStackTrace();
             } catch (IOException e) {
                 e.printStackTrace();
             } finally {
@@ -1139,34 +786,19 @@ public class DataRepository {
         return Collections.emptyList();
     }
 
-    public List<ModelObject> search(String pattern) {
-        List<ModelObject> result = new ArrayList<>();
-        search(getBooks(), pattern, result);
-        search(getBlogs(), pattern, result);
-        search(getCompanies(), pattern, result);
-        search(getPeople(), pattern, result);
-        search(getLibraries(), pattern, result);
-        search(getRealWorldApps(), pattern, result);
-        search(getTools(), pattern, result);
-        search(getVideos(), pattern, result);
-        search(getNews(), pattern, result);
-        search(getDownloads(), pattern, result);
-        search(getTutorials(), pattern, result);
-        search(getTips(), pattern, result);
-        search(getIkonliPacks(), pattern, result);
-        search(getMembers(), pattern, result);
-        search(getDocumentation(), pattern, result);
-        search(getLearnJavaFX(), pattern, result);
-        search(getLearnMobile(), pattern, result);
-        search(getLearnRaspberryPi(), pattern, result);
-        return result;
-    }
+    public File getRepositoryDirectory() {
+        if (testing) {
+            try {
+                String path = this.getClass().getProtectionDomain().getCodeSource().getLocation().toURI().getPath();
 
-    private void search(List<? extends ModelObject> modelObjects, String pattern, List<ModelObject> result) {
-        modelObjects.forEach(mo -> {
-            if (mo.matches(pattern)) {
-                result.add(mo);
+                // Convert the class file location to the Java file location
+                File classFile = new File(path, this.getClass().getName().replace('.', '/') + ".class");
+                return classFile.getParentFile().getParentFile().getParentFile().getParentFile().getParentFile().getParentFile().getParentFile();
+            } catch (URISyntaxException e) {
+                LOG.severe("Can not define source root path: " + e.getMessage());
+                return new File(System.getProperty("user.dir"));
             }
-        });
+        }
+        return REPO_DIRECTORY;
     }
 }
