+++ src/test/java/com/dlsc/jfxcentral/data/ImageManagerTest.java
@@ -32,14 +32,12 @@ public class ImageManagerTest {
         for (Handler handler : rootLogger.getHandlers()) {
             handler.setLevel(Level.FINE);
         }
-
-        DataRepository2.setTesting(true);
     }
 
     @Test
     public void shouldGetPersonImage() {
         // when .. then
-        DataRepository2.getInstance().getPeople().forEach(item -> {
+        DataRepository.getInstance().getPeople().forEach(item -> {
             ObjectProperty<Image> property = ImageManager.getInstance().personImageProperty(item);
             assertNotNull(property, "image property is null for item ID " + item.getId());
             assertNotNull(property.get(), "image is missing for item ID " + item.getId());
@@ -49,7 +47,7 @@ public class ImageManagerTest {
     @Test
     public void shouldGetNewsBanner() {
         // when .. then
-        DataRepository2.getInstance().getNews().forEach(item -> {
+        DataRepository.getInstance().getNews().forEach(item -> {
             ObjectProperty<Image> property = ImageManager.getInstance().newsBannerImageProperty(item);
             assertNotNull(property, "image property is null for item ID " + item.getId());
             assertNotNull(property.get(), "image is missing for item ID " + item.getId());
@@ -59,7 +57,7 @@ public class ImageManagerTest {
     @Test
     public void shouldGetDownloadBanner() {
         // when .. then
-        DataRepository2.getInstance().getDownloads().forEach(item -> {
+        DataRepository.getInstance().getDownloads().forEach(item -> {
             ObjectProperty<Image> property = ImageManager.getInstance().downloadBannerImageProperty(item);
             assertNotNull(property, "image property is null for item ID " + item.getId());
             assertNotNull(property.get(), "image is missing for item ID " + item.getId());
@@ -69,7 +67,7 @@ public class ImageManagerTest {
     @Test
     public void shouldGetTipBanner() {
         // when .. then
-        DataRepository2.getInstance().getTips().forEach(item -> {
+        DataRepository.getInstance().getTips().forEach(item -> {
             ObjectProperty<Image> property = ImageManager.getInstance().tipBannerImageProperty(item);
             assertNotNull(property, "image property is null for item ID " + item.getId());
             assertNotNull(property.get(), "image is missing for item ID " + item.getId());
@@ -79,7 +77,7 @@ public class ImageManagerTest {
     @Test
     public void shouldGetBlogPageImage() {
         // when .. then
-        DataRepository2.getInstance().getBlogs().forEach(item -> {
+        DataRepository.getInstance().getBlogs().forEach(item -> {
             ObjectProperty<Image> property = ImageManager.getInstance().blogPageImageProperty(item);
             assertNotNull(property, "image property is null for item ID " + item.getId());
             assertNotNull(property.get(), "image is missing for item ID " + item.getId());
@@ -89,7 +87,7 @@ public class ImageManagerTest {
     @Test
     public void shouldGetBlogIconImage() {
         // when .. then
-        DataRepository2.getInstance().getBlogs().forEach(item -> {
+        DataRepository.getInstance().getBlogs().forEach(item -> {
             ObjectProperty<Image> property = ImageManager.getInstance().blogIconImageProperty(item);
             assertNotNull(property, "image property is null for item ID " + item.getId());
             assertNotNull(property.get(), "image is missing for item ID " + item.getId());
@@ -99,7 +97,7 @@ public class ImageManagerTest {
     @Test
     public void shouldGetBlogPageLargeImage() {
         // when .. then
-        DataRepository2.getInstance().getBlogs().forEach(item -> {
+        DataRepository.getInstance().getBlogs().forEach(item -> {
             ObjectProperty<Image> property = ImageManager.getInstance().blogPageLargeImageProperty(item);
             assertNotNull(property, "image property is null for item ID " + item.getId());
             assertNotNull(property.get(), "image is missing for item ID " + item.getId());
@@ -109,7 +107,7 @@ public class ImageManagerTest {
     @Test
     public void shouldGetTutorialPageImage() {
         // when .. then
-        DataRepository2.getInstance().getTutorials().forEach(item -> {
+        DataRepository.getInstance().getTutorials().forEach(item -> {
             ObjectProperty<Image> property = ImageManager.getInstance().tutorialImageProperty(item);
             assertNotNull(property, "image property is null for item ID " + item.getId());
             assertNotNull(property.get(), "image is missing for item ID " + item.getId());
@@ -119,7 +117,7 @@ public class ImageManagerTest {
     @Test
     public void shouldGetTutorialPageLargeImage() {
         // when .. then
-        DataRepository2.getInstance().getTutorials().forEach(item -> {
+        DataRepository.getInstance().getTutorials().forEach(item -> {
             ObjectProperty<Image> property = ImageManager.getInstance().tutorialImageLargeProperty(item);
             assertNotNull(property, "image property is null for item ID " + item.getId());
             assertNotNull(property.get(), "image is missing for item ID " + item.getId());
@@ -129,7 +127,7 @@ public class ImageManagerTest {
     @Test
     public void shouldGetRealWorldImage() {
         // when .. then
-        DataRepository2.getInstance().getRealWorldApps().forEach(item -> {
+        DataRepository.getInstance().getRealWorldApps().forEach(item -> {
             ObjectProperty<Image> property = ImageManager.getInstance().realWorldAppImageProperty(item);
             assertNotNull(property, "small image property is null for item ID " + item.getId());
             assertNotNull(property.get(), "small image is missing for item ID " + item.getId());
@@ -139,7 +137,7 @@ public class ImageManagerTest {
     @Test
     public void shouldGetRealWorldImageLarge() {
         // when .. then
-        DataRepository2.getInstance().getRealWorldApps().forEach(item -> {
+        DataRepository.getInstance().getRealWorldApps().forEach(item -> {
             ObjectProperty<Image> property = ImageManager.getInstance().realWorldAppLargeImageProperty(item);
             assertNotNull(property, "large image property is null for item ID " + item.getId());
             assertNotNull(property.get(), "large image is missing for item ID " + item.getId());
@@ -149,7 +147,7 @@ public class ImageManagerTest {
     @Test
     public void shouldGetRealWorldBanner() {
         // when .. then
-        DataRepository2.getInstance().getRealWorldApps().forEach(item -> {
+        DataRepository.getInstance().getRealWorldApps().forEach(item -> {
             ObjectProperty<Image> property = ImageManager.getInstance().realWorldAppBannerImageProperty(item);
             assertNotNull(property, "banner image property is null for item ID " + item.getId());
             assertNotNull(property.get(), "banner image is missing for item ID " + item.getId());
@@ -159,7 +157,7 @@ public class ImageManagerTest {
     @Test
     public void shouldGetToolImage() {
         // when .. then
-        DataRepository2.getInstance().getTools().forEach(item -> {
+        DataRepository.getInstance().getTools().forEach(item -> {
             ObjectProperty<Image> property = ImageManager.getInstance().toolImageProperty(item);
             assertNotNull(property, "image property is null for item ID " + item.getId());
             assertNotNull(property.get(), "image is missing for item ID " + item.getId());
@@ -169,7 +167,7 @@ public class ImageManagerTest {
     @Test
     public void shouldGetUtilityImage() {
         // when .. then
-        DataRepository2.getInstance().getUtilities().forEach(item -> {
+        DataRepository.getInstance().getUtilities().forEach(item -> {
             ObjectProperty<Image> property = ImageManager.getInstance().utilityImageProperty(item);
             assertNotNull(property, "image property is null for item ID " + item.getId());
             assertNotNull(property.get(), "image is missing for item ID " + item.getId());
@@ -179,7 +177,7 @@ public class ImageManagerTest {
     @Test
     public void shouldGetCompanyImage() {
         // when .. then
-        DataRepository2.getInstance().getCompanies().forEach(item -> {
+        DataRepository.getInstance().getCompanies().forEach(item -> {
             ObjectProperty<Image> property = ImageManager.getInstance().companyImageProperty(item);
             assertNotNull(property, "image property is null for item ID " + item.getId());
             assertNotNull(property.get(), "image is missing for item ID " + item.getId());
@@ -189,7 +187,7 @@ public class ImageManagerTest {
     @Test
     public void shouldGetBookCoverImage() {
         // when .. then
-        DataRepository2.getInstance().getBooks().forEach(item -> {
+        DataRepository.getInstance().getBooks().forEach(item -> {
             ObjectProperty<Image> property = ImageManager.getInstance().bookCoverImageProperty(item);
             assertNotNull(property, "image property is null for item ID " + item.getId());
             assertNotNull(property.get(), "image is missing for item ID " + item.getId());
@@ -199,7 +197,7 @@ public class ImageManagerTest {
     @Disabled
     public void shouldGetYouTubeImage() {
         // when .. then
-        DataRepository2.getInstance().getVideos().forEach(item -> {
+        DataRepository.getInstance().getVideos().forEach(item -> {
             System.out.println("Youtube image url: " + ImageManager.getInstance().youTubeImageURL(item));
             ObjectProperty<Image> property = ImageManager.getInstance().youTubeImageProperty(item);
             assertNotNull(property, "image property is null for item ID " + item.getId() + ", name = " + item.getName());
@@ -228,7 +226,7 @@ public class ImageManagerTest {
     @Test
     public void shouldGetDocumentationImage() {
         // when .. then
-        DataRepository2.getInstance().getDocumentation().forEach(item -> {
+        DataRepository.getInstance().getDocumentation().forEach(item -> {
             ObjectProperty<Image> property = ImageManager.getInstance().documentationImageProperty(item);
             assertNotNull(property, "image property is null for item ID " + item.getId());
             assertNotNull(property.get(), "image is missing for item ID " + item.getId());
@@ -238,7 +236,7 @@ public class ImageManagerTest {
     @Test
     public void shouldGetLibraryFeaturedImage() {
         // when .. then
-        DataRepository2.getInstance().getLibraries().forEach(item -> {
+        DataRepository.getInstance().getLibraries().forEach(item -> {
             ObjectProperty<Image> featuredImageProperty = ImageManager.getInstance().libraryFeaturedImageProperty(item);
             assertNotNull(featuredImageProperty, "image featuredImageProperty is null for item ID " + item.getId());
             // featuredImageProperty.get() Can be null.
